Update excel processing
