+++ app/(ExcelProcessing)/excel/page.jsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { Button, Input } from "antd";
 import { LinkOutlined } from "@ant-design/icons";
